rename server folder into src folder
