+++ index.ts
@@ -1,1 +1,1 @@
-export * from './server';
+export * from './src';
